Add different data exchanger classes
