+++ psydac/ddm/tests/test_cart_2d.py
@@ -7,7 +7,8 @@ def run_cart_2d( verbose=False , nprocs=None, reverse_axis=None):
 
     import numpy as np
     from mpi4py       import MPI
-    from psydac.ddm.cart import DomainDecomposition, CartDecomposition, CartDataExchanger
+    from psydac.ddm.cart import DomainDecomposition, CartDecomposition
+    from psydac.ddm.blocking_data_exchanger import BlockingCartDataExachanger
 
     #---------------------------------------------------------------------------
     # INPUT PARAMETERS
@@ -69,7 +70,7 @@ def run_cart_2d( verbose=False , nprocs=None, reverse_axis=None):
     e1,e2 = cart.ends
 
     # Create object in charge of exchanging data between subdomains
-    synchronizer = CartDataExchanger( cart, u.dtype, coeff_shape=[2] )
+    synchronizer = BlockingCartDataExachanger( cart, u.dtype, coeff_shape=[2] )
 
     # Print some info
     if rank == 0:
@@ -92,7 +93,8 @@ def run_cart_2d( verbose=False , nprocs=None, reverse_axis=None):
     u[p1:-p1,p2:-p2,:] = [[(i1,i2) for i2 in range(s2,e2+1)] for i1 in range(s1,e1+1)]
 
     # Update ghost regions
-    synchronizer.update_ghost_regions( u )
+    synchronizer.start_update_ghost_regions( array=u )
+    synchronizer.end_update_ghost_regions()
 
     #---------------------------------------------------------------------------
     # CHECK RESULTS
